Refactor test_FloatingMultiplication to use test class
